refactor: Group repository and datasource classes
