+++ app/src/main/java/com/haidoan/android/stren/core/datasource/remote/base/ExercisesRemoteDataSource.kt
@@ -1,12 +1,12 @@
-package com.haidoan.android.stren.core.datasource.remote
+package com.haidoan.android.stren.core.datasource.remote.base
 
 import com.google.firebase.firestore.Query
 import com.haidoan.android.stren.core.model.Exercise
 import com.haidoan.android.stren.core.model.ExerciseCategory
 import com.haidoan.android.stren.core.model.ExerciseFilterStandards
 import com.haidoan.android.stren.core.model.MuscleGroup
-import com.haidoan.android.stren.core.repository.DEFAULT_ITEM_COUNT_LIMIT
-import com.haidoan.android.stren.core.repository.QueryWrapper
+import com.haidoan.android.stren.core.repository.base.DEFAULT_ITEM_COUNT_LIMIT
+import com.haidoan.android.stren.core.repository.impl.QueryWrapper
 
 interface ExercisesRemoteDataSource {
     /**
