+++ app/src/main/java/com/haidoan/android/stren/core/datasource/remote/base/RoutinesRemoteDataSource.kt
@@ -1,12 +1,16 @@
-package com.haidoan.android.stren.core.repository
+package com.haidoan.android.stren.core.datasource.remote.base
 
 import com.haidoan.android.stren.core.model.Routine
 import kotlinx.coroutines.flow.Flow
 
-interface RoutinesRepository {
+interface RoutinesRemoteDataSource {
     suspend fun getRoutinesStreamByUserId(userId: String): Flow<List<Routine>>
+
     suspend fun getRoutinesByUserId(userId: String): List<Routine>
+
     suspend fun getRoutineById(userId: String, routineId: String): Routine
+
     suspend fun addRoutine(userId: String, routine: Routine): String
+
     suspend fun updateRoutine(userId: String, routine: Routine)
 }
