+++ app/src/main/java/com/haidoan/android/stren/core/datasource/remote/di/DataSourceModule.kt
@@ -2,7 +2,14 @@ package com.haidoan.android.stren.core.datasource.remote.di
 
 import com.haidoan.android.stren.core.datasource.*
 import com.haidoan.android.stren.core.datasource.remote.*
-import com.haidoan.android.stren.core.network.FoodRetrofitDataSource
+import com.haidoan.android.stren.core.datasource.remote.base.ExercisesRemoteDataSource
+import com.haidoan.android.stren.core.datasource.remote.base.FoodRemoteDataSource
+import com.haidoan.android.stren.core.datasource.remote.base.RoutinesRemoteDataSource
+import com.haidoan.android.stren.core.datasource.remote.base.WorkoutRemoteDataSource
+import com.haidoan.android.stren.core.datasource.remote.impl.ExercisesFirestoreDataSource
+import com.haidoan.android.stren.core.datasource.remote.impl.RoutinesFirestoreDataSource
+import com.haidoan.android.stren.core.datasource.remote.impl.WorkoutFirestoreDataSource
+import com.haidoan.android.stren.core.datasource.remote.model.FoodRetrofitDataSource
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
