+++ app/src/main/java/com/haidoan/android/stren/core/datasource/remote/fake/FakeExercisesRemoteDataSource.kt
@@ -1,12 +1,12 @@
 package com.haidoan.android.stren.core.datasource.remote.fake
 
 import com.google.firebase.firestore.Query
-import com.haidoan.android.stren.core.datasource.remote.ExercisesRemoteDataSource
+import com.haidoan.android.stren.core.datasource.remote.base.ExercisesRemoteDataSource
 import com.haidoan.android.stren.core.model.Exercise
 import com.haidoan.android.stren.core.model.ExerciseCategory
 import com.haidoan.android.stren.core.model.ExerciseFilterStandards
 import com.haidoan.android.stren.core.model.MuscleGroup
-import com.haidoan.android.stren.core.repository.QueryWrapper
+import com.haidoan.android.stren.core.repository.impl.QueryWrapper
 import com.haidoan.android.stren.core.testing.data.EXERCISE_CATEGORIES_TEST_DATA
 
 class FakeExercisesRemoteDataSource : ExercisesRemoteDataSource {
