+++ app/src/main/java/com/haidoan/android/stren/core/repository/base/WorkoutsRepository.kt
@@ -1,4 +1,4 @@
-package com.haidoan.android.stren.core.repository
+package com.haidoan.android.stren.core.repository.base
 
 import com.haidoan.android.stren.core.model.Workout
 import kotlinx.coroutines.flow.Flow
