+++ app/src/main/java/com/haidoan/android/stren/feat/nutrition/food/FoodViewModel.kt
@@ -4,7 +4,7 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import androidx.paging.cachedIn
-import com.haidoan.android.stren.core.repository.FoodRepository
+import com.haidoan.android.stren.core.repository.base.FoodRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.flow.MutableStateFlow
